Refactored code as a helper
