+++ packages/eslint-plugin-pf-codemods/test/rules/v5/dataList-updated-callback.js
@@ -26,7 +26,7 @@ ruleTester.run("dataList-updated-callback", rule, {
       output: `import { DataList } from '@patternfly/react-core'; <DataList onSelectDataListItem={(_event, id) => onSelect(id)} />;`,
       errors: [
         {
-          message: `The "onSelectDataListItem" prop for DataList has been updated to include the event parameter as its first parameter. "onSelectDataListItem" handlers may require an update.`,
+          message: `The "onSelectDataListItem" prop for DataList has been updated to include the "event" parameter as its first parameter. "onSelectDataListItem" handlers may require an update.`,
           type: "JSXOpeningElement",
         },
       ],
@@ -36,7 +36,7 @@ ruleTester.run("dataList-updated-callback", rule, {
       output: `import { DataList } from '@patternfly/react-core'; <DataList onSelectDataListItem={(_event, id) => onSelect(id)} />;`,
       errors: [
         {
-          message: `The "onSelectDataListItem" prop for DataList has been updated to include the event parameter as its first parameter. "onSelectDataListItem" handlers may require an update.`,
+          message: `The "onSelectDataListItem" prop for DataList has been updated to include the "event" parameter as its first parameter. "onSelectDataListItem" handlers may require an update.`,
           type: "JSXOpeningElement",
         },
       ],
@@ -46,7 +46,7 @@ ruleTester.run("dataList-updated-callback", rule, {
       output: `import { DataList } from '@patternfly/react-core'; const onSelect = (_event, id) => {}; <DataList onSelectDataListItem={onSelect} />;`,
       errors: [
         {
-          message: `The "onSelectDataListItem" prop for DataList has been updated to include the event parameter as its first parameter. "onSelectDataListItem" handlers may require an update.`,
+          message: `The "onSelectDataListItem" prop for DataList has been updated to include the "event" parameter as its first parameter. "onSelectDataListItem" handlers may require an update.`,
           type: "JSXOpeningElement",
         },
       ],
@@ -56,7 +56,7 @@ ruleTester.run("dataList-updated-callback", rule, {
       output: `import { DataList } from '@patternfly/react-core'; function onSelect(_event, id) {}; <DataList onSelectDataListItem={onSelect} />;`,
       errors: [
         {
-          message: `The "onSelectDataListItem" prop for DataList has been updated to include the event parameter as its first parameter. "onSelectDataListItem" handlers may require an update.`,
+          message: `The "onSelectDataListItem" prop for DataList has been updated to include the "event" parameter as its first parameter. "onSelectDataListItem" handlers may require an update.`,
           type: "JSXOpeningElement",
         },
       ],
@@ -66,7 +66,7 @@ ruleTester.run("dataList-updated-callback", rule, {
       output: `import { DataList } from '@patternfly/react-core'; <DataList onSelectDataListItem={this.onSelect} />;`,
       errors: [
         {
-          message: `The "onSelectDataListItem" prop for DataList has been updated to include the event parameter as its first parameter. "onSelectDataListItem" handlers may require an update.`,
+          message: `The "onSelectDataListItem" prop for DataList has been updated to include the "event" parameter as its first parameter. "onSelectDataListItem" handlers may require an update.`,
           type: "JSXOpeningElement",
         },
       ],
@@ -76,7 +76,7 @@ ruleTester.run("dataList-updated-callback", rule, {
       output: `import { DataList as PFDataList } from '@patternfly/react-core'; <PFDataList onSelectDataListItem={(_event, id) => onSelect(id)} />;`,
       errors: [
         {
-          message: `The "onSelectDataListItem" prop for DataList has been updated to include the event parameter as its first parameter. "onSelectDataListItem" handlers may require an update.`,
+          message: `The "onSelectDataListItem" prop for PFDataList has been updated to include the "event" parameter as its first parameter. "onSelectDataListItem" handlers may require an update.`,
           type: "JSXOpeningElement",
         },
       ],
